+++ tests/L10nTest.php
@@ -20,7 +20,31 @@ class L10nTest extends PHPUnit\Framework\TestCase
 
         $l10n = new L10n('de');
         $this->assertEquals($langde['admin'], $l10n->message('admin'));
+        $this->assertEquals(true, isset($l10n['admin']));
         $this->assertEquals($langde['admin'], $l10n['admin']);
         $this->assertEquals('Undefined message!', $l10n['bla bla']);
+    }
+
+    public function testDefaultLanguage()
+    {
+        $langen = L10n::loadMessages('en');
+
+        $l10n = new L10n('en');
+        $this->assertEquals($langen['admin'], $l10n->message('admin'));
+    }
+
+    public function testInvalidLanguage()
+    {
+        $langen = L10n::loadMessages('en');
+
+        $l10n = new L10n('na');
+        $this->assertEquals($langen['admin'], $l10n->message('admin'));
+    }
+
+    public function testEmptyMessage()
+    {
+        $l10n = new L10n('en');
+        $this->assertEquals('', $l10n->message(null));
+        $this->assertEquals('', $l10n->message(''));
     }
 }